:art: separate cli
