+++ cita-cli/src/cli/abi_command.rs
@@ -1,0 +1,203 @@
+use clap::{App, Arg, ArgMatches, SubCommand};
+use serde_json::{self, Value};
+
+use cita_tool::{decode_input, decode_logs, decode_params, encode_input, encode_params, remove_0x};
+use interactive::GlobalConfig;
+use printer::Printer;
+
+/// Ethereum abi sub command
+pub fn abi_command() -> App<'static, 'static> {
+    let param_arg = Arg::with_name("param")
+        .long("param")
+        .takes_value(true)
+        .multiple(true)
+        .allow_hyphen_values(true)
+        .number_of_values(2)
+        .help("Function parameters");
+    let no_lenient_flag = Arg::with_name("no-lenient")
+        .long("no-lenient")
+        .help("Don't allow short representation of input params");
+
+    App::new("ethabi")
+        .about("ABI operation, encode parameter, generate code based on abi and parameters")
+        .subcommand(
+            SubCommand::with_name("encode")
+                .subcommand(
+                    SubCommand::with_name("function")
+                        .arg(
+                            Arg::with_name("file")
+                                .required(true)
+                                .index(1)
+                                .help("ABI json file path"),
+                        )
+                        .arg(
+                            Arg::with_name("name")
+                                .required(true)
+                                .index(2)
+                                .help("function name"),
+                        )
+                        .arg(param_arg.clone().number_of_values(1).value_name("value"))
+                        .arg(no_lenient_flag.clone()),
+                )
+                .subcommand(
+                    SubCommand::with_name("params")
+                        .arg(param_arg.clone().value_names(&["type", "value"]))
+                        .arg(no_lenient_flag),
+                ),
+        )
+        .subcommand(
+            SubCommand::with_name("decode")
+                .subcommand(
+                    SubCommand::with_name("params")
+                        .arg(
+                            Arg::with_name("type")
+                                .long("type")
+                                .takes_value(true)
+                                .multiple(true)
+                                .help("Decode types"),
+                        )
+                        .arg(
+                            Arg::with_name("data")
+                                .long("data")
+                                .takes_value(true)
+                                .help("Decode data"),
+                        ),
+                )
+                .subcommand(
+                    SubCommand::with_name("function")
+                        .arg(
+                            Arg::with_name("file")
+                                .required(true)
+                                .index(1)
+                                .help("ABI json file path"),
+                        )
+                        .arg(
+                            Arg::with_name("name")
+                                .required(true)
+                                .index(2)
+                                .help("function name"),
+                        )
+                        .arg(
+                            Arg::with_name("data")
+                                .long("data")
+                                .required(true)
+                                .takes_value(true)
+                                .help("Decode data"),
+                        ),
+                )
+                .subcommand(
+                    SubCommand::with_name("log")
+                        .arg(
+                            Arg::with_name("file")
+                                .required(true)
+                                .index(1)
+                                .help("ABI json file path"),
+                        )
+                        .arg(
+                            Arg::with_name("event")
+                                .required(true)
+                                .index(2)
+                                .help("event name"),
+                        )
+                        .arg(param_arg.clone().number_of_values(1).value_name("topic"))
+                        .arg(
+                            Arg::with_name("data")
+                                .long("data")
+                                .required(true)
+                                .takes_value(true)
+                                .help("Decode data"),
+                        ),
+                ),
+        )
+}
+
+/// ABI processor
+pub fn abi_processor(
+    sub_matches: &ArgMatches,
+    printer: &Printer,
+    env_variable: &GlobalConfig,
+) -> Result<(), String> {
+    let is_color = !sub_matches.is_present("no-color") && env_variable.color();
+    match sub_matches.subcommand() {
+        ("encode", Some(em)) => match em.subcommand() {
+            ("function", Some(m)) => {
+                let file = m.value_of("file").unwrap();
+                let name = m.value_of("name").unwrap();
+                let lenient = !m.is_present("no-lenient");
+                let values: Vec<String> = match m.values_of("param") {
+                    None => Vec::new(),
+                    Some(param) => param.map(|s| s.to_owned()).collect::<Vec<String>>(),
+                };
+                let output =
+                    encode_input(file, name, &values, lenient).map_err(|err| format!("{}", err))?;
+                printer.println(&Value::String(output), is_color);
+            }
+            ("params", Some(m)) => {
+                let lenient = !m.is_present("no-lenient");
+                let mut types: Vec<String> = Vec::new();
+                let mut values: Vec<String> = Vec::new();
+                let mut param_iter = m.values_of("param")
+                    .ok_or_else(|| format!("Please give at least one parameter."))?
+                    .peekable();
+                while param_iter.peek().is_some() {
+                    types.push(param_iter.next().unwrap().to_owned());
+                    values.push(param_iter.next().unwrap().to_owned());
+                }
+                let output =
+                    encode_params(&types, &values, lenient).map_err(|err| format!("{}", err))?;
+                printer.println(&Value::String(output), is_color);
+            }
+            _ => {
+                return Err(em.usage().to_owned());
+            }
+        },
+        ("decode", Some(em)) => match em.subcommand() {
+            ("params", Some(m)) => {
+                let types: Vec<String> = m.values_of("type")
+                    .ok_or_else(|| format!("Please give at least one parameter."))?
+                    .map(|value| value.to_owned())
+                    .collect();
+                let data = remove_0x(m.value_of("data").unwrap());
+                let output = decode_params(&types, data)
+                    .map_err(|err| format!("{}", err))?
+                    .iter()
+                    .map(|value| serde_json::from_str(value).unwrap())
+                    .collect();
+                printer.println(&Value::Array(output), is_color);
+            }
+            ("function", Some(m)) => {
+                let file = m.value_of("file").unwrap();
+                let name = m.value_of("name").unwrap();
+                let values = m.value_of("data").unwrap();
+                let output = decode_input(file, name, values)
+                    .map_err(|err| format!("{}", err))?
+                    .iter()
+                    .map(|value| serde_json::from_str(value).unwrap())
+                    .collect();
+                printer.println(&Value::Array(output), is_color);
+            }
+            ("log", Some(m)) => {
+                let file = m.value_of("file").unwrap();
+                let event = m.value_of("event").unwrap();
+                let topic: Vec<String> = match m.values_of("param") {
+                    None => Vec::new(),
+                    Some(param) => param.map(|s| s.to_owned()).collect::<Vec<String>>(),
+                };
+                let data = m.value_of("data").unwrap();
+                let output = decode_logs(file, event, &topic, data)
+                    .map_err(|err| format!("{}", err))?
+                    .iter()
+                    .map(|value| serde_json::from_str(value).unwrap())
+                    .collect();
+                printer.println(&Value::Array(output), is_color);
+            }
+            _ => {
+                return Err(em.usage().to_owned());
+            }
+        },
+        _ => {
+            return Err(sub_matches.usage().to_owned());
+        }
+    }
+    Ok(())
+}
