+++ cita-cli/src/interactive.rs
@@ -17,9 +17,9 @@ use serde_json;
 use shell_words;
 
 use cli::{
-    abi_processor, amend_processor, build_interactive, contract_processor, key_processor,
-    parse_privkey, rpc_processor, search_processor, store_processor, transfer_processor,
-    tx_processor,
+    abi_processor, amend_processor, benchmark_processor, build_interactive, contract_processor,
+    key_processor, parse_privkey, rpc_processor, search_processor, store_processor,
+    transfer_processor, tx_processor,
 };
 use printer::Printer;
 
@@ -169,6 +169,9 @@ pub fn start(url: &str) -> io::Result<()> {
                         Ok(())
                     }
                     ("tx", Some(m)) => tx_processor(m, &printer, Some(url.as_str()), &env_variable),
+                    ("benchmark", Some(m)) => {
+                        benchmark_processor(m, &printer, Some(url.as_str()), &env_variable)
+                    }
                     ("exit", _) => {
                         if let Err(err) = interface.save_history(history_file) {
                             eprintln!("Save command history failed: {}", err);
